requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
etelemetry==0.3.1
filelock==3.18.0
FITX==1.0.1
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-text-splitters==0.3.8
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prov==2.1.1
puremagic==1.30
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.4
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zstandard==0.23.0
